fixed bugs

+++ models/follower.js
@@ -38,11 +38,11 @@ module.exports = function (sequelize, DataTypes) {
                 'allowNull': false
             },
             'createdAt': {
-                'type': DataTypes.INTEGER,
+                'type': DataTypes.DOUBLE,
                 'allowNull': true
             },
             'updatedAt': {
-                'type': DataTypes.INTEGER,
+                'type': DataTypes.DOUBLE,
                 'allowNull': true
             }
         }
